Move post comments components to `components/posts`
